+++ src/main/java/kitchenpos/application/table/TableService.java
@@ -7,8 +7,8 @@ import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.OrderTableRepository;
-import kitchenpos.dto.order.request.ChangeGuestNumberRequest;
-import kitchenpos.dto.order.request.EmptyOrderTableRequest;
+import kitchenpos.dto.table.request.ChangeGuestNumberRequest;
+import kitchenpos.dto.table.request.EmptyOrderTableRequest;
 import kitchenpos.dto.table.request.OrderTableCreateRequest;
 import kitchenpos.dto.table.response.OrderTableResponse;
 import org.springframework.stereotype.Service;
