+++ src/main/java/kitchenpos/ui/table/TableRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.ui.table;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.application.table.TableService;
-import kitchenpos.dto.order.request.ChangeGuestNumberRequest;
-import kitchenpos.dto.order.request.EmptyOrderTableRequest;
+import kitchenpos.dto.table.request.ChangeGuestNumberRequest;
+import kitchenpos.dto.table.request.EmptyOrderTableRequest;
 import kitchenpos.dto.table.request.OrderTableCreateRequest;
 import kitchenpos.dto.table.response.OrderTableResponse;
 import org.springframework.http.ResponseEntity;
