+++ src/test/java/kitchenpos/application/table/TableServiceTest.java
@@ -15,8 +15,8 @@ import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.OrderTables;
 import kitchenpos.domain.table.TableGroup;
-import kitchenpos.dto.order.request.ChangeGuestNumberRequest;
-import kitchenpos.dto.order.request.EmptyOrderTableRequest;
+import kitchenpos.dto.table.request.ChangeGuestNumberRequest;
+import kitchenpos.dto.table.request.EmptyOrderTableRequest;
 import kitchenpos.dto.table.request.OrderTableCreateRequest;
 import kitchenpos.dto.table.response.OrderTableResponse;
 import org.junit.jupiter.api.DisplayName;
